👷 Start server after /start endpoint called (#57)
Currently we have a scenario where when the new deployment starts, it's running in parallel with the old one until the healthcheck is complete and the new one is made public. A side effect of this is we temporarily have 2 simultaneous connections to mongo/telegram which could cause data corruption. We need to ensure that in production, the server isn't started immediately but only after it's received a start signal (via an endpoint) which occurs after the previous deployment has already shutdown.

Set this up in production as it's not needed for local dev

trello.com/c/DLHKNqDQ